# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 52428800
     */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AdminStatus {
  isAdmin: boolean;
}

export type LearningLevel = typeof LearningLevel[keyof typeof LearningLevel];


export const LearningLevel = {
  Beginner: 'Beginner',
  Intermediate: 'Intermediate',
  Advanced: 'Advanced',
  All: 'All',
} as const;

export interface Assignment {
  id: number;
  title: string;
  description: string;
  level: LearningLevel;
  /** @nullable */
  assigneeUserId?: string | null;
  /** @nullable */
  assigneeName?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  resourcePath?: string | null;
  /** @nullable */
  resourceName?: string | null;
  isPublished: boolean;
  createdAt: string;
}

export interface CreateAssignmentRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /** @maxLength 4000 */
  description: string;
  level: LearningLevel;
  /** @nullable */
  assigneeUserId?: string | null;
  /** @nullable */
  assigneeName?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  resourcePath?: string | null;
  /** @nullable */
  resourceName?: string | null;
  isPublished: boolean;
}

export interface AssignmentMember {
  id: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
}

export type ListAssignmentsParams = {
level: LearningLevel;
};