# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Storage
    description: 관리자 파일 업로드와 저장 파일 제공
  - name: Assignments
    description: 레벨별 학습 과제 등록과 조회
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [Storage]
      summary: 관리자용 파일 업로드 주소 발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: 업로드 주소가 발급됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: 잘못된 파일 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: 로그인이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: 관리자 권한이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/status:
    get:
      operationId: getAdminStatus
      tags: [Storage]
      summary: 현재 로그인 사용자의 관리자 권한 확인
      responses:
        "200":
          description: 관리자 권한 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatus"
        "401":
          description: 로그인이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assignments:
    get:
      operationId: listAssignments
      tags: [Assignments]
      summary: 현재 레벨에 배정된 공개 과제 조회
      parameters:
        - name: level
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/LearningLevel"
      responses:
        "200":
          description: 공개 과제 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
        "400":
          description: 잘못된 레벨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: 로그인이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/assignments:
    get:
      operationId: listAdminAssignments
      tags: [Assignments]
      summary: 관리자가 전체 과제 조회
      responses:
        "200":
          description: 전체 과제 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
        "401":
          description: 로그인이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: 관리자 권한이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/members:
    get:
      operationId: listAdminMembers
      tags: [Assignments]
      summary: 과제 배정을 위한 회원 목록 조회
      responses:
        "200":
          description: 회원 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignmentMember"
        "401":
          description: 로그인이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: 관리자 권한이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createAssignment
      tags: [Assignments]
      summary: 레벨별 과제 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: 과제가 등록됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "400":
          description: 잘못된 과제 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: 로그인이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: 관리자 권한이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/assignments/{assignmentId}:
    delete:
      operationId: deleteAssignment
      tags: [Assignments]
      summary: 등록한 과제 삭제
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 과제가 삭제됨
        "400":
          description: 잘못된 과제 번호
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: 로그인이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: 관리자 권한이 필요함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: 과제를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [Storage]
      summary: 업로드된 파일 제공
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 파일 데이터
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: 파일을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
          maximum: 52428800
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AdminStatus:
      type: object
      required: [isAdmin]
      properties:
        isAdmin:
          type: boolean
    LearningLevel:
      type: string
      enum: [Beginner, Intermediate, Advanced, All]
    Assignment:
      type: object
      required: [id, title, description, level, isPublished, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          $ref: "#/components/schemas/LearningLevel"
        assigneeUserId:
          type: [string, "null"]
        assigneeName:
          type: [string, "null"]
        dueDate:
          type: [string, "null"]
        resourcePath:
          type: [string, "null"]
        resourceName:
          type: [string, "null"]
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateAssignmentRequest:
      type: object
      required: [title, description, level, isPublished]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 4000
        level:
          $ref: "#/components/schemas/LearningLevel"
        assigneeUserId:
          type: [string, "null"]
        assigneeName:
          type: [string, "null"]
        dueDate:
          type: [string, "null"]
        resourcePath:
          type: [string, "null"]
        resourceName:
          type: [string, "null"]
        isPublished:
          type: boolean
    AssignmentMember:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: [string, "null"]