# lib/api-zod/src/generated/types/assignmentMember.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentMember {
  id: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
}