# lib/api-zod/src/generated/types/createAssignmentRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningLevel } from './learningLevel';

export interface CreateAssignmentRequest {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /** @maxLength 4000 */
  description: string;
  level: LearningLevel;
  /** @nullable */
  assigneeUserId?: string | null;
  /** @nullable */
  assigneeName?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  resourcePath?: string | null;
  /** @nullable */
  resourceName?: string | null;
  isPublished: boolean;
}