# lib/api-zod/src/generated/types/learningLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type LearningLevel = typeof LearningLevel[keyof typeof LearningLevel];


export const LearningLevel = {
  Beginner: 'Beginner',
  Intermediate: 'Intermediate',
  Advanced: 'Advanced',
  All: 'All',
} as const;